+++ src/resources/scivis/tube.rs
@@ -1205,7 +1205,7 @@ impl ViewportGpuResources {
         }
         let (texture_view, has_texture): (&wgpu::TextureView, u32) =
             if let Some(id) = item.texture_id {
-                if let Some(tex) = self.textures.get(id as usize) {
+                if let Some(tex) = self.textures.get(id) {
                     (&tex.view, 1)
                 } else {
                     (&self.fallback_lut_view, 0)
